feat: update fixes methodology, add a flow.js helper
