+++ projects/fixes-coins/index.js
@@ -1,5 +1,5 @@
 // Fixes Inscription Protocol - 𝔉rc20 Treasury Pool: https://fixes.world/
-const { post } = require("../helper/http");
+const { callCadence } = require("../helper/chain/flow");
 
 let queryTVLCode = `
 import LiquidStaking from 0xd6f80565193ad727
@@ -59,22 +59,9 @@ fun main(): UFix64 {
 }
 `;
 
-const queryCodeBase64 = Buffer.from(queryTVLCode, "utf-8").toString("base64");
-
 async function tvl() {
   try {
-    const response = await post(
-      "https://rest-mainnet.onflow.org/v1/scripts",
-      { script: queryCodeBase64 },
-      {
-        headers: { "content-type": "application/json" },
-      }
-    );
-    let resEncoded = response;
-    let resString = Buffer.from(resEncoded, "base64").toString("utf-8");
-    let resJson = JSON.parse(resString);
-    let flowTokenTVL = Number(resJson.value);
-
+    const flowTokenTVL = await callCadence(queryTVLCode, true);
     return { flow: flowTokenTVL };
   } catch (error) {
     throw new Error("Couln't query scripts of Fixes coins", error);
